% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComprehensivePipelineUtils-package.R
\docType{package}
\name{ComprehensivePipelineUtils-package}
\alias{ComprehensivePipelineUtils}
\alias{ComprehensivePipelineUtils-package}
\title{ComprehensivePipelineUtils: CPU--comprehensive bioinfo-analysis pipeline utils package}
\description{
To wrap some useful code tools.
}
\author{
\strong{Maintainer}: Cal Luo \email{carl614317043@gmail.com}

}
\keyword{internal}
