% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm_to_tpm.R
\name{fpkm_to_tpm}
\alias{fpkm_to_tpm}
\title{From FPKM to TPM}
\usage{
fpkm_to_tpm(data_list, is_logged = TRUE, get_logged = TRUE)
}
\arguments{
\item{data_list}{FPKM data list}

\item{is_logged}{whther FPKM data is in log(X + 1) unit}

\item{get_logged}{whther to get TPM data in log(X + 1) unit}
}
\value{
TPM data list
}
\description{
From FPKM to TPM
}
\examples{
NULL
}
