% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_type.R
\name{get_gene_type}
\alias{get_gene_type}
\title{return gene type form gene ensembl ID based genecode V22 annotation}
\usage{
get_gene_type(genes, input_type = "ID")
}
\arguments{
\item{genes}{Genes of interest}

\item{input_type}{Symbol or ensembl ID}
}
\value{
Gene types
}
\description{
return gene type form gene ensembl ID based genecode V22 annotation
}
\examples{
library(ComprehensivePipelineUtils)
get_gene_type(c("H19", "HOTAIR"))
}
