% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_2_Symbol.R
\name{ID_2_Symbol}
\alias{ID_2_Symbol}
\title{return gene symbol form gene ensembel ID based genecode V22 annotation}
\usage{
ID_2_Symbol(ID_list)
}
\arguments{
\item{ID_list}{}
}
\value{
res
}
\description{
return gene symbol form gene ensembel ID based genecode V22 annotation
}
\examples{
library(ComprehensivePipelineUtils)
ID_2_Symbol(c("ENSG00000162444", "ENSG00000130939"))
}
