% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_symbol.R
\name{id_to_symbol}
\alias{id_to_symbol}
\title{return gene symbol form gene ensembel ID based genecode V22 annotation}
\usage{
id_to_symbol(id_list)
}
\arguments{
\item{id_list}{gene ensemble id list}
}
\value{
Symbol of genes
}
\description{
return gene symbol form gene ensembel ID based genecode V22 annotation
}
\examples{
library(ComprehensivePipelineUtils)
id_to_symbol(c("ENSG00000162444", "ENSG00000130939"))
}
