% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Symbol2ID.R
\name{Symbol2ID}
\alias{Symbol2ID}
\title{return gene ensembel ID form gene symbol based genecode V22 annotation}
\usage{
Symbol2ID(Symbol_list)
}
\arguments{
\item{Symbol_list}{}
}
\value{
res
}
\description{
return gene ensembel ID form gene symbol based genecode V22 annotation
}
\examples{
library(ComprehensivePipelineUtils)
Symbol2ID(c("H19", "HOTAIR"))
}
