% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_sample_filter.R
\name{tcga_sample_filter}
\alias{tcga_sample_filter}
\title{Get tumor&normal samples from mixed TCGA data}
\usage{
tcga_sample_filter(data_list, method = "best")
}
\arguments{
\item{data_list}{list of TCGA datasets}

\item{method}{sample selection method}
}
\value{
A list of tumor and normal samples
}
\description{
Get tumor&normal samples from mixed TCGA data
}
\examples{
NULL
}
