% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tcga_xena.R
\name{read_tcga_xena}
\alias{read_tcga_xena}
\title{Read TCGA mRNA profile from Xena platform in a dir}
\usage{
read_tcga_xena(
  folder_path,
  file_pattern = "*.tsv",
  use_all_cores = FALSE,
  cores = 4,
  add_file_info = FALSE
)
}
\arguments{
\item{folder_path}{the dir path of TCGA data}

\item{file_pattern}{file name pattern of TCGA data}

\item{use_all_cores}{whether use full cores}

\item{cores}{cores number to use}

\item{add_file_info}{whether add file info to res_list}
}
\value{
list of data and file path
}
\description{
Read TCGA mRNA profile from Xena platform in a dir
}
\examples{
NULL
}
