% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_to_id.R
\name{symbol_to_id}
\alias{symbol_to_id}
\title{return gene ensembel ID form gene symbol based genecode V22 annotation}
\usage{
symbol_to_id(symbol_list)
}
\arguments{
\item{symbol_list}{gene symbol list}
}
\value{
Ensemble gene IDs
}
\description{
return gene ensembel ID form gene symbol based genecode V22 annotation
}
\examples{
library(ComprehensivePipelineUtils)
symbol_to_id(c("H19", "HOTAIR"))
}
